+++ cylib/Primitives/MRT/Cylinder_MRT.cs
@@ -62,7 +62,8 @@ namespace cylib
 
         public void Dispose()
         {
-            em.removeMRT(DrawMRT);
+            if (em != null)
+                em.removeMRT(DrawMRT);
         }
     }
 }
